Moved last few functions from Import to proper location
