+++ code/drasil-code/drasil-code.cabal
@@ -31,6 +31,7 @@ library
     , Language.Drasil.Code.Imperative.FunctionCalls
     , Language.Drasil.Code.Imperative.GenerateGOOL
     , Language.Drasil.Code.Imperative.Generator
+    , Language.Drasil.Code.Imperative.Helpers
     , Language.Drasil.Code.Imperative.Import
     , Language.Drasil.Code.Imperative.LanguageRenderer
     , Language.Drasil.Code.Imperative.LanguageRenderer.JavaRenderer
@@ -95,6 +96,7 @@ executable codegenTest
     , Language.Drasil.Code.Imperative.FunctionCalls
     , Language.Drasil.Code.Imperative.GenerateGOOL
     , Language.Drasil.Code.Imperative.Generator
+    , Language.Drasil.Code.Imperative.Helpers
     , Language.Drasil.Code.Imperative.Import
     , Language.Drasil.Code.Imperative.Logging
     , Language.Drasil.Code.Imperative.Modules
